Add guest speakers data and adjust UI to fit template
