+++ index.html
@@ -16,30 +16,43 @@
 <body class="flex">
   <main>
     <header>
-      <nav class="mobile-nav">
+      <div class="pre-nav sub-section-grey">
+        <ul><a href="#"><i class="fa fa-facebook"></i></a></ul>
+        <ul><a href="#"><i class="fa fa-twitter"></i></a></ul>
+        <ul><a class="sub-section-text-white" href="#">English</a></ul>
+        <ul><a class="sub-section-text-white" href="#">My Page</a></ul>
+        <ul><a class="sub-section-text-white" href="#">Logout</a></ul>
+      </div>
+      <nav class="mobile-nav flex">
         <div id="menu-btn" data-menu>
           <div class="menu-bar"></div>
           <div class="menu-bar"></div>
           <div class="menu-bar"></div>
         </div>
+        <a href="index.html"><img class="logo-mobile" src="images/logo2-removebg-preview.png" alt=""></a>
       </nav>
       <nav class="desktop-nav flex display-none">
         <a href="#"><img class="logo" src="images/logo2-removebg-preview.png" alt=""></a>
         <ul class="flex">
           <li><a href="index2.html">About</a></li>
           <li><a href="#program-section">Program</a></li>
-          <li><a href="#speakers-section">Featured Speakers</a></li>
-          <li><a href="#partners">Partners</a></li>
+          <li><a href="#speakers-section">Join</a></li>
+          <li><a href="#partners">Sponsors</a></li>
+          <li><a href="index2.html/#see-past-header">News</a></li>
+          <li><a href="#">
+            <div class="campaign">AFF campaign</div>
+          </a></li>
         </ul>
       </nav>
       <section class="headline-section">
-        <h1 class="main-title">WEST AFRICAN FICTION FESTIVAL</h1>
+        <p class="pre-header">"Hello! Fiction world"</p>
+        <h1 class="main-title">AFRICAN FICTION FESTIVAL 2023</h1>
         <p class="main-paragraph">
-          Lorem ipsum dolor sit amet, consectetur adipisicing elit. Mollitia, exercitationem adipisci quia maxime nostrum similique esse eaque ex a quibusdam dignissimos, repellendus odit optio saepe.
+          A yearly festival to remind us all of the roots and purpose of African literature, creating a positive change with people from all 55 nation states of the African continent is taking place in October, in Nigeria.
         </p>
         <div class="program-date">
           <p id="bold-text">2023.04.17(THU) ~ 23(WED)</p>
-          <p class="section-text-white">@ Eagle Square, Abuja, Nigeria</p>
+          <p class="section-text-grey">@ Eagle Square, Abuja, Nigeria</p>
         </div>
       </section>
     </header>
@@ -50,29 +63,39 @@
       </div>
       <div class="program-list flex flex-column">
         <div class="flex">
-          <img class="program-icons flex" src="images/program-icon-1.png" alt="program 1 icon">
+          <div class="program-icon-div">
+            <img class="program-icons flex" src="images/program-icon-1.png" alt="program 1 icon">
+          </div>
           <h3 class="sub-section-header-white flex align-center red-color">Lecture</h3>
-          <p class="program-description">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quas assumenda iure, sunt mollitia inventore beatae.</p>
+          <p class="program-description">Listen to the speakers from various countries talk about the hitory and prospects of African storytelling.</p>
         </div>
         <div class="flex">
-          <img class="program-icons flex" src="images/program-icon-2.png" alt="program 1 icon">
-          <h3 class="sub-section-header-white flex align-center red-color">Lecture</h3>
-          <p class="program-description">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quas assumenda iure, sunt mollitia inventore beatae.</p>
+          <div class="program-icon-div">
+            <img class="program-icons flex" src="images/program-icon-2.png" alt="program 1 icon">
+          </div>
+          <h3 class="sub-section-header-white flex align-center red-color">Auction</h3>
+          <p class="program-description">Appreciate various writing styles from writers across all Africa, organized by the West African Art Gallery, Accra</p>
         </div>
         <div class="flex">
-          <img class="program-icons flex" src="images/program-icon-3.png" alt="program 1 icon">
-          <h3 class="sub-section-header-white flex align-center red-color">Lecture</h3>
-          <p class="program-description">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quas assumenda iure, sunt mollitia inventore beatae.</p>
+          <div class="program-icon-div">
+            <img class="program-icons flex" src="images/program-icon-3.png" alt="program 1 icon">
+          </div>
+          <h3 class="sub-section-header-white flex align-center red-color">Connect</h3>
+          <p class="program-description">Have the time to share your thoughts and opinions with experts for each topic.</p>
         </div>
         <div class="flex">
-          <img class="program-icons flex" src="images/program-icon-4.png" alt="program 1 icon">
-          <h3 class="sub-section-header-white flex align-center red-color">Lecture</h3>
-          <p class="program-description">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quas assumenda iure, sunt mollitia inventore beatae.</p>
+          <div class="program-icon-div">
+            <img class="program-icons flex" src="images/program-icon-4.png" alt="program 1 icon">
+          </div>
+          <h3 class="sub-section-header-white flex align-center red-color">Workshop</h3>
+          <p class="program-description">Try writng your own piece using open source license rather than just watching others do it.</p>
         </div>
         <div class="flex">
-          <img class="program-icons flex" src="images/program-icon-5.png" alt="program 1 icon">
-          <h3 class="sub-section-header-white flex align-center red-color">Lecture</h3>
-          <p class="program-description">Lorem ipsum dolor, sit amet consectetur adipisicing elit. Quas assumenda iure, sunt mollitia inventore beatae.</p>
+          <div class="program-icon-div">
+            <img class="program-icons flex" src="images/program-icon-5.png" alt="program 1 icon">
+          </div>
+          <h3 class="sub-section-header-white flex align-center red-color">Tour</h3>
+          <p class="program-description">From the Famous Lagos up north to Kaduna and then southwards to Port Harcout, take your time to enjoy our country</p>
         </div>
       </div>
       <div class="join-btn flex">
