+++ index.js
@@ -1,7 +1,10 @@
+import { authors, seeMoreAuthors, fullAuthorList } from "./speakerData.js";
+
+console.log(authors)
+
 const main = document.querySelector('main');
 
 const hamburger = document.querySelector('[data-menu]');
-console.log(hamburger);
 
 const menu = () => {
   const modal = document.createElement('section');
@@ -13,13 +16,16 @@ const menu = () => {
   <li><a class='hyperlink' href="index2.html">About <div class="middle-underline flex justify-center">
     <div class="line"></div>
   </div></a></li>
-    <li><a class='hyperlink' href="#program-section">Program <div class="middle-underline flex justify-center">
+    <li><a class='hyperlink' href="index.html#program-section">Program <div class="middle-underline flex justify-center">
       <div class="line"></div>
     </div></a></li>
-    <li><a class='hyperlink' href="#speakers-section">Featured Speakers <div class="middle-underline flex justify-center">
+    <li><a class='hyperlink' href="index.html#speakers-section">Featured Speakers <div class="middle-underline flex justify-center">
       <div class="line"></div>
     </div></a></li>
-    <li><a class='hyperlink' href="#partners">Partners <div class="middle-underline flex justify-center">
+    <li><a class='hyperlink' href="index2.html#partners">Sponsors <div class="middle-underline flex justify-center">
+      <div class="line"></div>
+    </div></a></li>
+    <li><a class='hyperlink' href="index2.html#see-past-header">News <div class="middle-underline flex justify-center">
       <div class="line"></div>
     </div></a></li>
   </ul>`;
@@ -43,128 +49,12 @@ const menu = () => {
 
 hamburger.addEventListener('click', menu);
 
-const authors = [
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/images.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Chimamanda Ngozi Adichie',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/chimamanda.jpeg',
-      alt: ''
-    }
-  }
-];
 
-const seeMoreAuthors = [
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Thiongo.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Imbolo-Mbue.webp',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/mbue2.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Amma-darko.jpeg',
-      alt: ''
-    }
-  }
-];
-
-const fullAuthorList = [
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/images.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Chimamanda Ngozi Adichie',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/chimamanda.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Thiongo.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Imbolo-Mbue.webp',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/mbue2.jpeg',
-      alt: ''
-    }
-  },
-  {
-    name: 'Wole Soyinka',
-    honours: 'Nobel Laurette 1975',
-    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
-    image: {
-      src: 'images/Amma-darko.jpeg',
-      alt: ''
-    }
-  }
-];
 
 const cardsContainer = document.querySelector('.cards-container');
 
 const guestSpeakers = (arr) => {
   for (let i = 0; i < arr.length; i += 1) {
-    // const promise = new Promise()
     const speakerDetails = arr[i];
     console.log(speakerDetails);
     const card = document.createElement('div');
@@ -184,6 +74,10 @@ const guestSpeakers = (arr) => {
   return cardsContainer;
 }
 guestSpeakers(authors);
+
+
+// setTimeout(function() {guestSpeakers(authors)}, 3000);
+
 
 const seeMore = document.createElement('div');
 
@@ -209,7 +103,7 @@ function guestSpeakersDesktop(arr) {
     const card = document.createElement('div');
     card.className = 'cards';
     card.classList.add('desktop-cards');
-    card.innerHTML = `<img class="speaker-image" src=${speakerDetails.image.src}" alt="Speaker's Image"/>
+    card.innerHTML = `<img class="speaker-image" src=${speakerDetails.image.src}" alt=${speakerDetails.name}/>
     <div class="speaker-details">
       <h3 class="sub-section-header-black">${speakerDetails.name}</h3>
       <p class="sub-section-text-black red-color"><i>${speakerDetails.honours}</i></p>
