+++ speakerData.js
@@ -1,0 +1,237 @@
+// const authors = [
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/images.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Chimamanda Ngozi Adichie',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/chimamanda.jpeg',
+//       alt: ''
+//     }
+//   }
+// ];
+
+
+// const seeMoreAuthors = [
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Thiongo.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Imbolo-Mbue.webp',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/mbue2.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Amma-darko.jpeg',
+//       alt: ''
+//     }
+//   }
+// ];
+
+// const fullAuthorList = [
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/images.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Chimamanda Ngozi Adichie',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/chimamanda.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Thiongo.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Imbolo-Mbue.webp',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/mbue2.jpeg',
+//       alt: ''
+//     }
+//   },
+//   {
+//     name: 'Wole Soyinka',
+//     honours: 'Nobel Laurette 1975',
+//     books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+//     image: {
+//       src: 'images/Amma-darko.jpeg',
+//       alt: ''
+//     }
+//   }
+// ];
+
+
+const authors = [
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://guardian.ng/wp-content/uploads/2022/10/Chimamanda-e1667045283649-600x448.jpg'
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://guardian.ng/wp-content/uploads/2022/10/Chimamanda-e1667045283649-600x448.jpg'
+    }
+  }
+];
+
+
+
+const seeMoreAuthors = [
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://static.dw.com/image/16245728_902.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  }
+];
+
+const fullAuthorList = [
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://guardian.ng/wp-content/uploads/2022/10/Chimamanda-e1667045283649-600x448.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Chimamanda Ngozi Adichie',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  },
+  {
+    name: 'Wole Soyinka',
+    honours: 'Nobel Laurette 1975',
+    books: 'Notable Works: Purple Hibiscus, The Dead Zone, Break in the woods, Cry alone, Half of the Yellow sun, Blood is not the Answer, War is the Problem, Live and Let us Live',
+    image: {
+      src: 'https://gazettengr.com/wp-content/uploads/Wole-Soyinka2-671x400.jpg',
+      alt: ''
+    }
+  }
+];
+
+export { authors, seeMoreAuthors, fullAuthorList };
